refactor: Give metrics a lobotomy and remove AverageMeters
