+++ pytorch_msssim/__init__.py
@@ -4,7 +4,6 @@ from math import exp
 import numpy as np
 
 
-
 def gaussian(window_size, sigma):
     gauss = torch.Tensor([exp(-(x - window_size//2)**2/float(2*sigma**2)) for x in range(window_size)])
     return gauss/gauss.sum()
